core/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pydantic>=2.12.4",
    "rerun-sdk>=0.27.2",
    "rich>=14.2.0",
    "tyro>=0.9.35",
    "webpolicy",
]

[tool.uv.sources]
webpolicy = { git = "https://github.com/mhyatt000/webpolicy" }
